+++ src/bacen_analysis/exceptions.py
@@ -127,3 +127,4 @@ class AmbiguousIdentifierError(BacenAnalysisError):
         
         super().__init__(msg)
 
+